backend/requirements.txt
fastapi
uvicorn
firebase-admin
openai
python-dotenv
aiohttp